+++ __check__/SequentialContainerTest/Deque.test.ts
@@ -9,8 +9,8 @@ for (let i = 0; i < testNum; ++i) {
 }
 
 function judgeSequentialContainer(
-  container: SequentialContainer<number, unknown>,
-  myVector: SequentialContainer<number, unknown>
+  container: SequentialContainer<number>,
+  myVector: SequentialContainer<number>
 ) {
   expect(container.size()).toEqual(myVector.size());
   container.forEach((element, index) => {